requirements.txt
openai==0.27.2
discord.py==2.2.2
python-dotenv==1.0.0
tiktoken==0.3.2